efls-data/requirements.txt
scipy==1.5.4
numpy==1.16.4
tensorflow==1.14.0
murmurhash
pylint
grpcio-tools
etcd3
kubernetes
rsa
psutil
sqlalchemy==1.2.19
prison==0.1.3
apache-flink==1.13.1
mmh3~=3.0.0
cardinality~=0.1.1
gmpy2
cityhash